orchestrator/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.21.0
